# backend_web/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "backend_web, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/id": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/by-type": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/add-product": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "caption",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "price",
            "type": "number",
            "format": "double"
          },
          {
            "in": "formData",
            "name": "type",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "PHnum",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "images",
            "type": "array",
            "items": {
              "format": "binary",
              "type": "string"
            }
          },
          {
            "in": "formData",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/{productId}": {
      "put": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "caption",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "price",
            "type": "number",
            "format": "double"
          },
          {
            "in": "formData",
            "name": "type",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "PHnum",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "images",
            "type": "array",
            "items": {
              "format": "binary",
              "type": "string"
            }
          },
          {
            "in": "formData",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/{id}": {
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/mdlUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/id": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/mdlUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetLikedProducts/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/ToggleLike": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddLikeDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AddLikeDto": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "DateOnly": {
      "type": "object",
      "properties": {
        "year": {
          "format": "int32",
          "type": "integer"
        },
        "month": {
          "format": "int32",
          "type": "integer"
        },
        "day": {
          "format": "int32",
          "type": "integer"
        },
        "dayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "dayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      },
      "additionalProperties": false
    },
    "DayOfWeek": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "$ref": "#/definitions/DateOnly"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "mdlUser": {
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 100,
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "maxLength": 15,
          "minLength": 8,
          "pattern": "^(?=.*[a-zA-Z])(?=.*\\d)[a-zA-Z\\d]{8,15}$",
          "type": "string"
        },
        "phone": {
          "pattern": "^(077|078|079)\\d{7}$",
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}